// src/entities/form/model/formValuesWithRenovation.ts
export type FormValuesWithRenovationType = {
    address: string;
    houseType: string;
    roofType: string;
    rooms: string;
    square: string;
    floor: string;
    floorTotal: string;
    hasLift: boolean;
    parkingOnRoof: boolean;
    parkingOnGround: boolean;
    parkingUnderGround: boolean;
    renovation: string;
};